Implement injecting profiles into existing config
When the user is in a repo directory, they can press enter on a selected
profile opening a prompt. This enables the user the inject the selected
profile data into the local config of the repo meaning, their username
and email used for commits is changed to that of the selected profile.